#586: Extract isValidEntryPoint to a function.
